Streamline UI (#34)
* Fixes issue #23 whereby intermediate step of loading and image and
then calculating are combined into new view controller which does
everything. Cleaned up display of image data and moved some of it to Debug view.

Check in code but still need to to test and tweak a little bit more.

* Added override for traitCollectionDidChange in all view controllers so
that we can better hanble dark/light mode transitions.  Plus,
centralize htmlToAttributedString to main ViewController so that we
can handle setting foreground color so that dark and light modes work
with textView widgets.

* Added ActivityIndicator to LoadViewCalculate when a DEM is being downloaded.
Set version to 2.5.  Feature branch complete.  Fixes #23